Add constraint for roles

+++ Software/src/main/java/controllers/RolesUpdateController.java
@@ -21,18 +21,15 @@ import org.hibernate.SessionFactory;
 import repositories.PermissionRepository;
 import repositories.RolesDetailRepository;
 import utils.*;
-import validation.CustomerValidation;
 import validation.RolesValidation;
 
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Objects;
 import java.util.ResourceBundle;
 import java.util.stream.Collectors;
 
 public class RolesUpdateController implements Initializable {
-
     @FXML
     private AnchorPane host;
     @FXML
@@ -70,14 +67,12 @@ public class RolesUpdateController implements Initializable {
         // Add value to permissionTable
         List<Permissions> permissionsList = PermissionRepository.getAll(session);
         TableHelper.setPermissionNameTable(permissionsList, permissionTable, nameCol);
-
         // Add value to selectedPermissionList
         List<RolesDetail> rolesDetailList = RolesDetailRepository.getByRolesId(roles.getId());
         for (RolesDetail item : rolesDetailList) {
-            selectedPermission.add(permissionsList.stream().filter(t -> t.getCode().equals(item.getPermissionCode())).findFirst().get().getName());
+            selectedPermission.add(permissionsList.stream().filter(t -> t.getCode().equals(item.getPermissions().getCode())).findFirst().get().getName());
         }
         populateSelectedPermissionList();
-
         // Set Roles name
         nameHolder.setText(roles.getName());
     }
@@ -96,32 +91,33 @@ public class RolesUpdateController implements Initializable {
     @FXML
     void delete(ActionEvent event) {
         SessionFactory factory = HibernateUtils.getSessionFactory();
-        Session session = factory.getCurrentSession();
+        Session session;
 
-        // Delete RolesDetail
-        RolesDetailRepository.deleteByRoleId(roles.getId());
+        if (RolesValidation.validateDelete(factory, roles).size() == 0) {
+            // Delete RolesDetail
+            session = factory.openSession();
+            RolesDetailRepository.deleteByRoleId(session, roles.getId());
+            // Delete Roles
+            session = factory.openSession();
+            session.beginTransaction();
+            session.delete(roles);
+            session.getTransaction().commit();
 
-        // Delete Roles
-        session = factory.openSession();
-        session.beginTransaction();
-        session.delete(roles);
-        session.getTransaction().commit();
+            // Refresh content table
+            RolesCategoryController.getInstance().refresh();
+            // clear rolesHolder
+            rolesHolder.setRoles(null);
+            // Close stage
+            StageHelper.closeStage(event);
+            // Show alert box
+            AlertBoxHelper.showMessageBox("Xoá thành công");
+            // Unhide host
+            AnchorPane host = MainNavigatorController.instance.getHost();
+            host.setDisable(false);
+        } else {
+            errorMessage.setText(RolesValidation.validateDelete(factory, roles).get(0));
+        }
 
-        // Refresh content table
-        RolesCategoryController.getInstance().refresh();
-
-        // clear rolesHolder
-        rolesHolder.setRoles(null);
-
-        // Close stage
-        StageHelper.closeStage(event);
-
-        // Show alert box
-        AlertBoxHelper.showMessageBox("Xoá thành công");
-
-        // Unhide host
-        AnchorPane host = MainNavigatorController.instance.getHost();
-        host.setDisable(false);
     }
 
     @FXML
@@ -147,47 +143,42 @@ public class RolesUpdateController implements Initializable {
 
     @FXML
     void update(ActionEvent event) {
-        // Create session
         SessionFactory factory = HibernateUtils.getSessionFactory();
         Session session = factory.getCurrentSession();
-        session.beginTransaction();
 
         Roles roles = rolesHolder.getRoles();
         roles.setName(nameHolder.getText());
-
-        List<String> validateUpdate = RolesValidation.validateUpdate(session, roles);
+        List<String> validateUpdate = RolesValidation.validateUpdate(factory, roles);
         if (validateUpdate.size() == 0) {
             // Update Roles Info
             session = factory.openSession();
             session.beginTransaction();
             session.saveOrUpdate(roles);
             session.getTransaction().commit();
-
             // Update Roles Detail
-            RolesDetailRepository.deleteByRoleId(roles.getId());
-            session.beginTransaction();
+            session = factory.openSession();
+            RolesDetailRepository.deleteByRoleId(session, roles.getId());
             for (String item : selectedPermission){
+                session = factory.openSession();
                 Permissions permissions = PermissionRepository.getByName(session, item);
+
                 RolesDetail rolesDetail = new RolesDetail();
                 rolesDetail.setId(UUIDHelper.generateType4UUID().toString());
-                rolesDetail.setRoleId(roles.getId());
-                rolesDetail.setPermissionCode(permissions.getCode());
+                rolesDetail.setRoles(roles);
+                rolesDetail.setPermissions(permissions);
+                session = factory.openSession();
+                session.beginTransaction();
                 session.save(rolesDetail);
+                session.getTransaction().commit();
             }
-            session.getTransaction().commit();
-
             // Close stage
             StageHelper.closeStage(event);
-
             // Show alert box
             AlertBoxHelper.showMessageBox("Cập nhật thành công");
-
             // Refresh content table
             RolesCategoryController.getInstance().refresh();
-
             // Set roles holder
             rolesHolder.setRoles(roles);
-
             // Unhide host
             AnchorPane host = MainNavigatorController.instance.getHost();
             host.setDisable(false);
